feat: 将 weex v2 逻辑抽到单独的文件
